requirements.txt
requests
dnspython
colorama
urllib3
tldextract
